[add] 環境変数などを設定するスクリプトを追加

+++ dotfilesLink.sh
@@ -1,19 +1,18 @@
 #!/bin/sh
 set -u
 
-# 今のディレクトリ
-# dotfilesディレクトリに移動する
-BASEDIR=$(dirname $0)
-cd $BASEDIR
-
 # dotfilesディレクトリにある、ドットから始まり2文字以上の名前のファイルに対して
 for f in .??*; do
-	[ "$f" = ".git" ] && continue
-	[ "$f" = ".gitconfig.local.template" ] && continue
-	[ "$f" = ".gitmodules" ] && continue
+  [ "$f" = ".git" ] && continue
+  [ "$f" = ".gitconfig.local.template" ] && continue
+  [ "$f" = ".gitmodules" ] && continue
+  [ `echo "$f" | grep 'brew'` ] && continue
+
   if [ -d ~/"$f" ]; then
     rm -rf ~/"$f"
   fi
-	# シンボリックリンクを貼る
-	ln -snfv ${PWD}/"$f" ~/
+
+  # シンボリックリンクを貼る
+  ln -snfv ${PWD}/"$f" ~/
 done
+
